Android expt:added fragment support, plus included clean up/restruture
