+++ build_android/res/layout/main.xml
@@ -33,30 +33,10 @@
 			android:layout_width="match_parent"
 			android:layout_height="match_parent" >
 			<LinearLayout
-				android:orientation="vertical"
-				android:layout_width="fill_parent"
-				android:layout_height="fill_parent">
-
-				<ScrollView
-					android:id="@+id/game_scroller"
-					android:layout_width="match_parent"
-					android:layout_height="match_parent"
-					android:layout_weight="1"
-					android:scrollbars="none"
-					android:fillViewport="true" >
-
-					<TableLayout
-						android:id="@+id/game_grid"
-						android:layout_width="match_parent"
-						android:layout_height="wrap_content"
-						android:layout_marginTop="10dp"
-						android:paddingBottom="10dp"
-						android:stretchColumns="*" >
-
-						<requestFocus />
-					</TableLayout>
-				</ScrollView>
-			</LinearLayout>
+				android:id="@+id/container"
+				android:layout_width="match_parent"
+				android:layout_height="match_parent"
+				android:orientation="horizontal" />
 		</FrameLayout>
 		<!-- The navigation drawer -->
 		<fragment android:id="@+id/navigation_drawer"
