+++ build_android/res/values/strings.xml
@@ -28,4 +28,6 @@
     <string name="navigation_drawer_open">Open navigation drawer</string>
     <string name="navigation_drawer_close">Close navigation drawer</string>
 	<string name="pressback_exit">Please click BACK again to exit</string>
+
+    <string name="external_launcher">PLAY - PS2 Game Launcher</string>
 </resources>
